requirements.txt
python-telegram-bot==11.1.0
py-money
requests